fix: 필수교양과목 본문 컴포넌트 분리
